WIP: dual scatterplot
