+++ packages/constants-internal/src/coordination.ts
@@ -75,6 +75,49 @@ export const COMPONENT_COORDINATION_TYPES = {
     CoordinationType.CONTOUR_COLOR_ENCODING,
     CoordinationType.CONTOUR_COLOR,
   ],
+  [ViewType.DUAL_SCATTERPLOT]: [
+    CoordinationType.DATASET,
+    CoordinationType.OBS_TYPE,
+    CoordinationType.FEATURE_TYPE,
+    CoordinationType.FEATURE_VALUE_TYPE,
+    CoordinationType.OBS_LABELS_TYPE,
+    CoordinationType.EMBEDDING_TYPE,
+    CoordinationType.EMBEDDING_ZOOM,
+    CoordinationType.EMBEDDING_ROTATION,
+    CoordinationType.EMBEDDING_TARGET_X,
+    CoordinationType.EMBEDDING_TARGET_Y,
+    CoordinationType.EMBEDDING_TARGET_Z,
+    CoordinationType.EMBEDDING_OBS_SET_POLYGONS_VISIBLE,
+    CoordinationType.EMBEDDING_OBS_SET_LABELS_VISIBLE,
+    CoordinationType.EMBEDDING_OBS_SET_LABEL_SIZE,
+    CoordinationType.EMBEDDING_OBS_RADIUS,
+    CoordinationType.EMBEDDING_OBS_RADIUS_MODE,
+    CoordinationType.EMBEDDING_OBS_OPACITY,
+    CoordinationType.EMBEDDING_OBS_OPACITY_MODE,
+    CoordinationType.OBS_FILTER,
+    CoordinationType.OBS_HIGHLIGHT,
+    CoordinationType.OBS_SET_SELECTION,
+    CoordinationType.OBS_SET_FILTER,
+    CoordinationType.OBS_SET_HIGHLIGHT,
+    CoordinationType.OBS_SET_COLOR,
+    CoordinationType.FEATURE_HIGHLIGHT,
+    CoordinationType.FEATURE_SELECTION,
+    CoordinationType.FEATURE_VALUE_COLORMAP,
+    CoordinationType.FEATURE_VALUE_COLORMAP_RANGE,
+    CoordinationType.OBS_COLOR_ENCODING,
+    CoordinationType.ADDITIONAL_OBS_SETS,
+    CoordinationType.TOOLTIPS_VISIBLE,
+    CoordinationType.SAMPLE_TYPE,
+    CoordinationType.SAMPLE_SET_SELECTION,
+    CoordinationType.SAMPLE_SET_FILTER,
+    CoordinationType.SAMPLE_SET_COLOR,
+    CoordinationType.EMBEDDING_POINTS_VISIBLE,
+    CoordinationType.EMBEDDING_CONTOURS_VISIBLE,
+    CoordinationType.EMBEDDING_CONTOURS_FILLED,
+    CoordinationType.EMBEDDING_CONTOUR_PERCENTILES,
+    CoordinationType.CONTOUR_COLOR_ENCODING,
+    CoordinationType.CONTOUR_COLOR,
+  ],
   [ViewType.GATING]: [
     CoordinationType.DATASET,
     CoordinationType.OBS_TYPE,
