+++ packages/main/all/src/base-plugins.ts
@@ -71,7 +71,10 @@ import {
 // Register view type plugins
 import { DescriptionSubscriber } from '@vitessce/description';
 import { ObsSetsManagerSubscriber } from '@vitessce/obs-sets-manager';
-import { EmbeddingScatterplotSubscriber } from '@vitessce/scatterplot-embedding';
+import {
+  EmbeddingScatterplotSubscriber,
+  DualEmbeddingScatterplotSubscriber,
+} from '@vitessce/scatterplot-embedding';
 import { GatingSubscriber } from '@vitessce/scatterplot-gating';
 import { SpatialSubscriber } from '@vitessce/spatial';
 import { SpatialBetaSubscriber } from '@vitessce/spatial-beta';
@@ -221,6 +224,7 @@ export const baseViewTypes = [
   makeViewType(ViewType.DESCRIPTION, DescriptionSubscriber),
   makeViewType(ViewType.OBS_SETS, ObsSetsManagerSubscriber),
   makeViewType(ViewType.SCATTERPLOT, EmbeddingScatterplotSubscriber),
+  makeViewType(ViewType.DUAL_SCATTERPLOT, DualEmbeddingScatterplotSubscriber),
   makeViewType(ViewType.GATING, GatingSubscriber),
   makeViewType(ViewType.SPATIAL, SpatialSubscriber),
   makeViewType(ViewType.SPATIAL_BETA, SpatialBetaSubscriber),
