+++ packages/view-types/scatterplot-embedding/src/index.js
@@ -1,1 +1,2 @@
 export { EmbeddingScatterplotSubscriber } from './EmbeddingScatterplotSubscriber.js';
+export { DualEmbeddingScatterplotSubscriber } from './DualEmbeddingScatterplotSubscriber.js';
